+++ share/pot/common.sh
@@ -168,6 +168,10 @@ _conf_check()
 		_qerror "$1" "POT_FS_ROOT is mandatory"
 		return 1 # false
 	fi
+	if ! getent group "${POT_GROUP:-pot}" >/dev/null 2>&1; then
+		_qerror "$1" "Group '${POT_GROUP:-pot}' is missing, create it or change POT_GROUP"
+		return 1 # false
+	fi
 	return 0 # true
 }
 
@@ -231,6 +235,16 @@ _zfs_exist()
 	_mnt_="$(zfs list -H -o mountpoint "$1" 2> /dev/null )"
 	if [ "$_mnt_" != "$2" ]; then
 		return 1 # false
+	fi
+	return 0 # true
+}
+
+# check if the dataset $1 is mounted
+# $1 the dataset NAME
+_zfs_mounted()
+{
+	if [ "$(zfs get -Ho value mounted "$1")" != "yes" ]; then
+		return 1; # false
 	fi
 	return 0 # true
 }
@@ -1076,8 +1090,20 @@ pot-cmd()
 	shift
 	if [ ! -r "${_POT_INCLUDE}/${_cmd}.sh" ]; then
 		_error "Fatal error! $_cmd implementation not found!"
-		exit 1
+		${EXIT} 1
 	fi
+
+	if [ "$_cmd" != "init" ]&& [ "$_cmd" != "de-init" ] ; then
+		if [ ! -d "$POT_FS_ROOT" ]; then
+			_error "$POT_FS_ROOT does not exist, please run 'pot init'"
+			${EXIT} 1
+		fi
+		if [ ! -r "$POT_FS_ROOT" ]; then
+			_error "Current user has no read access to $POT_FS_ROOT"
+			${EXIT} 1
+		fi
+	fi
+
 	# shellcheck disable=SC1090
 	. "${_POT_INCLUDE}/${_cmd}.sh"
 	_func=pot-${_cmd}